requirements.txt
sentencepiece
beautifulsoup4
tqdm
pdfplumber
transformers
huggingface_hub
numpy
pandas
scikit-learn
torch
gluonnlp
mxnet
setuptools
wheel
urllib3
PyYAML
six
git+https://github.com/SKTBrain/KoBERT.git#egg=kobert_tokenizer&subdirectory=kobert_hf
git+https://github.com/haven-jeon/PyKoSpacing.git
google-genai
openai
labelbox
opencv_python
shapely
pyproj
pyjeotile
psycopg2-binary
label-studio
matplotlib
pyahocorasick